data sizes and offsets: uint32_t -> uint64_t

+++ src/db/file.h
@@ -8,9 +8,9 @@ typedef struct {
 	uint32_t blocks_size; // size of working+draft
 	uint32_t draft_blocks_size; // size of draft
 	uint32_t blocks_capacity;
-	uint32_t data_size; // working+draft
+	uint64_t data_size; // working+draft
 	uint32_t headers_offset;
-	uint32_t data_offset;
+	uint64_t data_offset;
 } Metadata;
 
 typedef struct {
@@ -35,8 +35,8 @@ typedef struct {
 	uint32_t block_unique_id;
 	Entity_type type;
 	Block_status status;
-	uint32_t data_offset;
-	uint32_t data_size;
+	uint64_t data_offset;
+	uint64_t data_size;
 } Header_block;
 
 typedef struct {
